--- packages/ra-core/src/sideEffect/saga/index.js
@@ -1,10 +1,0 @@
-export auth from './auth';
-export cancelSaga from './cancelSaga';
-export crudFetch from './crudFetch';
-export crudResponse from './crudResponse';
-export crudSaga from './crudSaga';
-export notificationSaga from './notificationSaga';
-export redirectionSaga from './redirectionSaga';
-export referenceFetch from './referenceFetch';
-export refreshSaga from './refreshSaga';
-export i18n from './i18n';
